organized the code a bit for better structure
