+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import Navbar from "./components/navbar/Navbar";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import BottomNavbar from "./components/navbar/bottomNavbar";
-import Home from "./components/navbar/Home";
+import Home from "./pages/Home";
 
 function App() {
   let newDate = new Date()
