+++ src/components/Home.js
@@ -5,8 +5,7 @@ import VideoSection from '../../../videos/VideoSection';
 function Home(){
     return (
         <>
-        
-        <VideoSection/>
+            <VideoSection/>
         </>
     )
 }
