--- src/components/navbar/bottomNavbar.css
@@ -1,24 +1,0 @@
-li {
-    padding: 14px;
-}
-
-#page-title {
-    font-size: 2rem;
-    font-style: italic;
-}
-
-.bottom-navbar {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-    background-color: beige;
-}
-
-.bottom-navbar li {
-    list-style: none;
-    display:inline-block;
-    margin:auto;
-    }
-
-
-
