+++ src/components/navbar/navbar.scss
@@ -1,0 +1,44 @@
+@use "../../scss/colors" as *;
+@use "../../scss/font" as *;
+
+#page-title {
+    font-size: 2rem;
+    font-style: italic;
+}
+
+.navbar {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    background-color: beige;
+
+    .left-navbar {
+        float: left;
+    }
+    
+    .right-navbar {
+        float: right;
+    }
+    li {
+        list-style: none;
+        display:inline-block;
+        margin:auto;
+        padding: 14px;
+    }
+}
+
+.bottom-navbar {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    background-color: beige;
+
+    .left-bottomNavbar {
+        padding: 14px;
+        list-style: none;
+        display:inline-block;
+        margin:auto;
+    }
+}
+
+
