+++ src/components/video/VideoSecion.css
@@ -4,7 +4,6 @@ video{
     height: 100%;
     position: fixed 100%;
     z-index: -4;
-
 }
 
 #textOverlay{
@@ -13,8 +12,6 @@ video{
     left: 32%;
     z-index: 1;
     color:#fff;
-    
-    
 }
 
 .video-container{
@@ -27,15 +24,12 @@ video{
     align-items: center;
     box-shadow: inset 0 0 0 1000px rgba(0,0,0,0.2);
     object-fit: contain;
-   
-
 }
 
 .video-container >h1{
     color: #fff;
     font-size: 100px;
-    margin-top: -100px;
-     
+    margin-top: -100px;    
 }
 
 .video-container >p{
