+++ src/scss/_colors.scss
@@ -1,0 +1,2 @@
+$backgroundcolor: white;
+
