+++ src/scss/_font.scss
@@ -1,0 +1,3 @@
+@import url('https://fonts.googleapis.com/css2?family=Heebo&display=swap');
+
+$header-font: 'Heebo', sans-serif;
